Reorganising and working calibrated dependance
