+++ til_base_model/param/demo/dependance_prevalence_all_years.csv
@@ -1,0 +1,52 @@
+PERIOD,DEPENDANTS
+200901,1122004
+201001,1157171
+201101,1186781
+201201,1213266
+201301,1238063
+201401,1261881
+201501,1284919
+201601,1307066
+201701,1328061
+201801,1347628
+201901,1365573
+202001,1381853
+202101,1396619
+202201,1410224
+202301,1423210
+202401,1436276
+202501,1450217
+202601,1465850
+202701,1483931
+202801,1505085
+202901,1529757
+203001,1558166
+203101,1590295
+203201,1625874
+203301,1664406
+203401,1705196
+203501,1747410
+203601,1790149
+203701,1832523
+203801,1873713
+203901,1913023
+204001,1949907
+204101,1983995
+204201,2015087
+204301,2043156
+204401,2068320
+204501,2090818
+204601,2110969
+204701,2129129
+204801,2145652
+204901,2160861
+205001,2175021
+205101,2188325
+205201,2200880
+205301,2212709
+205401,2223759
+205501,2233923
+205601,2243072
+205701,2251088
+205801,2257922
+205901,2263649
