+++ til_base_model/tests/test_destinie.py
@@ -1,7 +1,8 @@
 # -*- coding:utf-8 -*-
 
 
-from til_base_model.tests.base import create_til_simulation, plot_population2, plot_csv, population_diagnostic
+from til_base_model.tests.base import create_til_simulation
+
 
 simulation = create_til_simulation(capitalized_name = 'Destinie', uniform_weight = 1922.721)
